# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageHealth": {
        "enum": [
          "critical",
          "serious",
          "fair",
          "good"
        ],
        "title": "PageHealth",
        "type": "string"
      },
      "PageMetricsRead": {
        "properties": {
          "critical_violation_count": {
            "title": "Critical Violation Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "page_health": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageHealth"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "violation_count": {
            "title": "Violation Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "url",
          "page_health",
          "violation_count",
          "critical_violation_count"
        ],
        "title": "PageMetricsRead",
        "type": "object"
      },
      "PageResultRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "inapplicable_count": {
            "title": "Inapplicable Count",
            "type": "integer"
          },
          "page_health": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageHealth"
              },
              {
                "type": "null"
              }
            ]
          },
          "passes_count": {
            "title": "Passes Count",
            "type": "integer"
          },
          "scan_run_id": {
            "title": "Scan Run Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scan_run_id",
          "url",
          "page_health",
          "passes_count",
          "inapplicable_count",
          "created_at",
          "updated_at"
        ],
        "title": "PageResultRead",
        "type": "object"
      },
...
      "Page_PageMetricsRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PageMetricsRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[PageMetricsRead]",
        "type": "object"
      },
...
      "Page_ScanRunRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScanRunRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ScanRunRead]",
        "type": "object"
      },
...
      "ScanRunCreate": {
        "properties": {
          "scanned_at": {
            "format": "date-time",
            "title": "Scanned At",
            "type": "string"
          }
        },
        "required": [
          "scanned_at"
        ],
        "title": "ScanRunCreate",
        "type": "object"
      },
      "ScanRunRead": {
        "properties": {
          "app_id": {
            "title": "App Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "scanned_at": {
            "format": "date-time",
            "title": "Scanned At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ScanRunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "app_id",
          "status",
          "scanned_at",
          "created_at",
          "updated_at"
        ],
        "title": "ScanRunRead",
        "type": "object"
      },
      "ScanRunStatus": {
        "enum": [
          "pending",
          "completed"
        ],
        "title": "ScanRunStatus",
        "type": "string"
      },
      "ScanRunStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ScanRunStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ScanRunStatusUpdate",
        "type": "object"
      },
      "ScanRunSummaryRead": {
        "properties": {
          "avg_violations_per_page": {
            "title": "Avg Violations Per Page",
            "type": "number"
          },
          "pages_with_critical_violations": {
            "title": "Pages With Critical Violations",
            "type": "integer"
          },
          "pages_with_violations": {
            "title": "Pages With Violations",
            "type": "integer"
          },
          "pct_pages_with_critical_violations": {
            "title": "Pct Pages With Critical Violations",
            "type": "number"
          },
          "pct_pages_with_violations": {
            "title": "Pct Pages With Violations",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          },
          "total_violations": {
            "title": "Total Violations",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "total_violations",
          "pages_with_violations",
          "pages_with_critical_violations",
          "total_pages",
          "avg_violations_per_page",
          "pct_pages_with_violations",
          "pct_pages_with_critical_violations"
        ],
        "title": "ScanRunSummaryRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Accessibility Health API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/apps/{app_id}/scan-runs": {
      "get": {
        "operationId": "list_scan_runs",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ScanRunRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scan Runs",
        "tags": [
          "scan-runs"
        ]
      },
      "post": {
        "operationId": "create_scan_run",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scan Run",
        "tags": [
          "scan-runs"
        ]
      }
    },
...
    "/api/v1/scan-runs/{scan_run_id}": {
      "delete": {
        "operationId": "delete_scan_run",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scan Run",
        "tags": [
          "scan-runs"
        ]
      },
      "get": {
        "operationId": "get_scan_run",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan Run",
        "tags": [
          "scan-runs"
        ]
      },
      "patch": {
        "operationId": "update_scan_run_status",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRunStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scan Run Status",
        "tags": [
          "scan-runs"
        ]
      }
    },
...
    "/api/v1/scan-runs/{scan_run_id}/pages": {
      "get": {
        "operationId": "list_scan_run_pages",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PageMetricsRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scan Run Pages",
        "tags": [
          "scan-runs"
        ]
      },
      "post": {
        "operationId": "create_page_result",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Raw Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Page Result",
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/scan-runs/{scan_run_id}/summary": {
      "get": {
        "operationId": "get_scan_run_summary",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunSummaryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan Run Summary",
        "tags": [
          "scan-runs"
        ]
      }
    }
  }
}